Moved models to separate file
